+++ cmd/get/get_pps.go
@@ -75,16 +75,12 @@ func getPps(cmd *cobra.Command, args []string) error {
 		params.Set("pageSize", strconv.Itoa(ppsPageSize))
 	}
 
-	// Filter systems based on system IDs provided via the root-level --system-ids flag.
-	// If no system IDs are provided, this will return all systems for the active profile.
 	systems, err = config.FilterSystems(config.Data, config.ActiveProfileName, config.SystemIDs)
 	if err != nil {
 		return err
 	}
 
-	// Loop through the filtered systems and get PPS data for each one.
 	for _, system = range systems {
-		// Define the endpoint for getting PPS data for the current system.
 		endpoint = fmt.Sprintf("%s/api/systems/%d/ports-protocols", config.Data.URL, system.ID)
 
 		if len(params) > 0 {
@@ -96,7 +92,6 @@ func getPps(cmd *cobra.Command, args []string) error {
 			return err
 		}
 
-		// Print the response in the specified output format.
 		err = output.Response(response, config.OutputFormat)
 		if err != nil {
 			return err
@@ -107,7 +102,6 @@ func getPps(cmd *cobra.Command, args []string) error {
 }
 
 func init() {
-	// Define flags for the "emu get pps" subcommand.
 	getPpsCmd.PersistentFlags().IntVarP(&ppsPageIndex, "page-index", "", ppsPageIndexDefault, "Page index")
 	getPpsCmd.PersistentFlags().IntVarP(&ppsPageIndex, "page-size", "", ppsPageSizeDefault, "Page size")
 }