+++ cmd/get/get_workflows.go
@@ -49,25 +49,20 @@ func getWorkflows(cmd *cobra.Command, args []string) error {
 		systems  []models.System
 	)
 
-	// Filter systems based on system IDs provided via the root-level --system-ids flag.
-	// If no system IDs are provided, this will return all systems for the active profile.
 	systems, err = config.FilterSystems(config.Data, config.ActiveProfileName, config.SystemIDs)
 	if err != nil {
 		return err
 	}
 
-	// Loop through the filtered systems and get workflow data for each one.
 	for _, system = range systems {
-		// Define the endpoint for getting workflow data for the current system.
+
 		endpoint = fmt.Sprintf("%s/api/systems/%d/approval/pac", config.Data.URL, system.ID)
 
-		// Make the request for workflow data.
 		response, err = emass.Get(system.ConfigProfile, endpoint)
 		if err != nil {
 			return err
 		}
 
-		// Print the response in the specified output format.
 		err = output.Response(response, config.OutputFormat)
 		if err != nil {
 			return err